KA-207 Refactor NoteRecord and merging of NoteRecords
